pyproject.toml
[tool.poetry]
name = "reflex-project"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
reflex = "^0.5.4"
reflex-webcam = "^0.0.5"
opencv-python = "^4.10.0.84"
pillow = "^10.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
